Move nested sort tests into a separate file
